even more refactoring, part 2
