+++ mehari/src/db/create/mod.rs
@@ -5,26 +5,21 @@ use crate::common::trace_rss_now;
 use crate::pbs::txs::{SourceVersion, TxSeqDatabase};
 use anyhow::{Error, anyhow};
 use cli::Args;
-use enumflags2::{BitFlag, BitFlags};
 use hgvs::data::cdot::json::models as cdot_models;
-use hgvs::data::cdot::json::models::{BioType, Gene, Tag, Transcript};
 use itertools::Itertools;
-use models::{GeneId, Identifier, LabelEntry, ReportEntry};
-use once_cell::sync::Lazy;
+use models::{LabelEntry, ReportEntry};
 use prost::Message;
 use serde_json::json;
 use serde_with::DisplayFromStr;
-use std::cmp::{PartialEq, Reverse};
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufWriter;
 use std::path::Path;
 use std::{io::Write, time::Instant};
-use strum::Display;
 use thousands::Separable;
 
 mod build;
-mod cdot;
+pub(crate) mod cdot;
 pub mod cli;
 mod filter;
 mod gff3;
@@ -39,553 +34,6 @@ use crate::db::create::filter::{
 };
 use crate::db::create::gff3::load_gff3;
 use models::*;
-
-/// Mitochondrial accessions.
-const MITOCHONDRIAL_ACCESSIONS: &[&str] = &["NC_012920.1", "NC_001807.4"];
-
-trait TranscriptExt {
-    fn cds_length(&self) -> Option<u32>;
-
-    fn protein_coding(&self) -> bool;
-
-    fn is_on_contig(&self, contig: &str) -> bool;
-}
-impl TranscriptExt for Transcript {
-    fn cds_length(&self) -> Option<u32> {
-        self.start_codon
-            .and_then(|start| self.stop_codon.map(|stop| stop.abs_diff(start)))
-    }
-
-    fn protein_coding(&self) -> bool {
-        matches!(self.protein.as_ref().map(|p| !p.is_empty()), Some(true))
-            || self
-                .biotype
-                .as_ref()
-                .is_some_and(|bt| bt.iter().any(|b| b.is_protein_coding()))
-    }
-
-    fn is_on_contig(&self, contig: &str) -> bool {
-        self.genome_builds.values().any(|gb| gb.contig == contig)
-    }
-}
-
-trait BioTypeExt {
-    fn is_protein_coding(&self) -> bool;
-}
-
-impl BioTypeExt for BioType {
-    fn is_protein_coding(&self) -> bool {
-        // see https://www.ensembl.org/info/genome/genebuild/biotypes.html
-        // and https://www.ensembl.org/Help/Faq?id=468
-        matches!(self, BioType::ProteinCoding | BioType::MRna)
-    }
-}
-
-static DISCARD_BIOTYPES_TRANSCRIPTS: Lazy<HashSet<BioType>> = Lazy::new(|| {
-    HashSet::from([
-        // BioType::PseudogenicTranscript,
-        BioType::IgCGene,
-        BioType::IgDGene,
-        BioType::IgJGene,
-        BioType::IgVGene,
-        BioType::TrCGene,
-        BioType::TrDGene,
-        BioType::TrJGene,
-        BioType::TrVGene,
-        BioType::AberrantProcessedTranscript,
-        BioType::UnconfirmedTranscript,
-        BioType::NmdTranscriptVariant,
-    ])
-});
-
-// this used to contain `BioType::Pseudogene`,
-// but we keep pseudogenes (as there are cases where a protein still is produced in certain individuals)
-static DISCARD_BIOTYPES_GENES: Lazy<HashSet<BioType>> = Lazy::new(|| {
-    HashSet::from([
-        BioType::CGeneSegment,
-        BioType::DGeneSegment,
-        BioType::JGeneSegment,
-        BioType::VGeneSegment,
-    ])
-});
-
-#[derive(Debug, Clone, Default)]
-struct TranscriptLoader {
-    genome_release: String,
-    annotation_version: String,
-    transcript_id_to_transcript: HashMap<TranscriptId, Transcript>,
-    gene_id_to_gene: HashMap<GeneId, Gene>,
-    gene_id_to_transcript_ids: HashMap<GeneId, Vec<TranscriptId>>,
-    discards: HashMap<Identifier, BitFlags<Reason>>,
-    fixes: HashMap<Identifier, BitFlags<Fix>>,
-    disable_filters: bool,
-}
-
-impl TranscriptLoader {
-    fn new(genome_release: String, disable_filters: bool) -> Self {
-        Self {
-            genome_release,
-            disable_filters,
-            ..Default::default()
-        }
-    }
-
-    fn merge(&mut self, other: &mut Self) -> &mut Self {
-        assert_eq!(self.genome_release, other.genome_release);
-        self.disable_filters |= other.disable_filters;
-
-        // Use annotation_version from whichever provides it, preferring non-empty
-        if self.annotation_version.is_empty() {
-            self.annotation_version
-                .clone_from(&other.annotation_version);
-        }
-
-        for (gene_id, gene) in other.gene_id_to_gene.drain() {
-            if let Some(old_gene) = self.gene_id_to_gene.insert(gene_id.clone(), gene.clone()) {
-                tracing::warn!(
-                    "Overwriting gene: {}\n{:?},\n{:?}",
-                    &gene_id,
-                    &old_gene,
-                    &gene
-                );
-            }
-        }
-
-        for (tx_id, tx) in other.transcript_id_to_transcript.drain() {
-            if let Some(old_tx) = self.transcript_id_to_transcript.insert(tx_id, tx.clone()) {
-                tracing::warn!(
-                    "Overwriting transcript: {}\n{:?},\n{:?}",
-                    old_tx.id,
-                    &old_tx,
-                    &tx
-                );
-            }
-        }
-
-        for (gene_id, mut tx_ids) in other.gene_id_to_transcript_ids.drain() {
-            let ids = self.gene_id_to_transcript_ids.entry(gene_id).or_default();
-            ids.append(&mut tx_ids);
-            ids.sort_unstable();
-            ids.dedup();
-        }
-
-        for (id, reason) in other.discards.drain() {
-            *self.discards.entry(id).or_default() |= reason;
-        }
-        for (id, fix) in other.fixes.drain() {
-            *self.fixes.entry(id).or_default() |= fix;
-        }
-        self
-    }
-
-    fn apply_fixes(
-        &mut self,
-        transcript_id_to_tags: &Option<HashMap<TranscriptId, Vec<FeatureTag>>>,
-    ) {
-        self.fix_transcript_genome_builds();
-        if let Some(txid_to_label) = transcript_id_to_tags {
-            self.update_transcript_tags(txid_to_label);
-        }
-        self.fix_cds();
-    }
-
-    fn update_pseudogene_status(&mut self) -> Result<(), Error> {
-        let empty = Reason::empty();
-        for (gene_id, gene) in &self.gene_id_to_gene {
-            if !self
-                .discards
-                .get(&Identifier::Gene(gene_id.clone()))
-                .unwrap_or(&empty)
-                .contains(Reason::NoTranscriptLeft)
-            {
-                continue;
-            }
-            if gene
-                .biotype
-                .as_ref()
-                .is_some_and(|bt| bt.contains(&BioType::Pseudogene))
-            {
-                *self
-                    .discards
-                    .entry(Identifier::Gene(gene_id.clone()))
-                    .or_default() |= Reason::Pseudogene;
-            }
-        }
-        Ok(())
-    }
-
-    fn gather_transcript_stats(&self) -> Result<(usize, usize, usize), Error> {
-        let (mut n_mane_select, mut n_mane_plus_clinical, mut n_mt) = (0, 0, 0);
-        for tx in self.transcript_id_to_transcript.values() {
-            let mut is_mane_select = false;
-            let mut is_mane_plus_clinical = false;
-            for gb in tx.genome_builds.values() {
-                if MITOCHONDRIAL_ACCESSIONS.contains(&gb.contig.as_str()) {
-                    n_mt += 1;
-                }
-                if let Some(tag) = &gb.tag {
-                    if tag.contains(&Tag::ManeSelect) {
-                        is_mane_select = true;
-                    }
-                    if tag.contains(&Tag::ManePlusClinical) {
-                        is_mane_plus_clinical = true;
-                    }
-                }
-            }
-            if is_mane_select {
-                n_mane_select += 1;
-            }
-            if is_mane_plus_clinical {
-                n_mane_plus_clinical += 1;
-            }
-        }
-        Ok((n_mt, n_mane_select, n_mane_plus_clinical))
-    }
-
-    fn update_transcript_tags(
-        &mut self,
-        transcript_id_to_tags: &HashMap<TranscriptId, Vec<FeatureTag>>,
-    ) {
-        self.transcript_id_to_transcript
-            .iter_mut()
-            .for_each(|(tx_id, tx)| {
-                // transfer MANE-related labels from TSV file
-                let tx_id_no_version = tx_id.without_version().unwrap();
-                if let Some(tags) = transcript_id_to_tags.get(&tx_id_no_version) {
-                    let mut any_change = false;
-                    tx.genome_builds.iter_mut().for_each(|(_, alignment)| {
-                        if let Some(alignment_tag) = &mut alignment.tag {
-                            let tags_to_add: Vec<Tag> = tags
-                                .iter()
-                                .filter_map(|t| {
-                                    let standard = Tag::from(t.clone());
-                                    if alignment_tag.contains(&standard) {
-                                        None
-                                    } else {
-                                        let backported = Tag::from(t.to_backported());
-                                        if alignment_tag.contains(&backported) {
-                                            None
-                                        } else {
-                                            Some(backported)
-                                        }
-                                    }
-                                })
-                                .collect();
-
-                            if !tags_to_add.is_empty() {
-                                alignment_tag.extend(tags_to_add);
-                                alignment_tag.sort();
-                                alignment_tag.dedup();
-                                any_change = true;
-                            }
-                        }
-                    });
-
-                    if any_change {
-                        self.fixes
-                            .entry(Identifier::Transcript(tx_id.clone()))
-                            .or_default()
-                            .insert(Fix::Tags);
-                    }
-                }
-            });
-    }
-
-    fn fix_transcript_genome_builds(&mut self) {
-        let genome_release = self.genome_release.clone();
-        self.transcript_id_to_transcript
-            .values_mut()
-            .for_each(|tx| {
-                let n = tx.genome_builds.len();
-                tx.genome_builds
-                    .retain(|key, _| key.eq_ignore_ascii_case(&genome_release));
-                if n != tx.genome_builds.len() {
-                    self.fixes
-                        .entry(Identifier::Transcript(
-                            TranscriptId::try_new(&tx.id).unwrap(),
-                        ))
-                        .or_default()
-                        .insert(Fix::GenomeBuild);
-                }
-            });
-    }
-
-    fn fix_cds(&mut self) {
-        self.transcript_id_to_transcript
-            .values_mut()
-            .filter(|tx| tx.protein_coding())
-            .filter_map(|tx| {
-                tx.start_codon
-                    .and_then(|start| tx.stop_codon.map(|stop| (start, stop)))
-                    .and_then(|(cds_start, cds_end)| {
-                        let cds_len = cds_end - cds_start;
-                        (cds_len % 3 != 0).then_some((tx, cds_start, cds_end, cds_len))
-                    })
-            })
-            .for_each(|(tx, _cds_start, cds_end, cds_len)| {
-                assert_eq!(tx.genome_builds.len(), 1);
-                for gb in tx.genome_builds.values_mut() {
-                    let delta = 3 - (cds_len % 3);
-                    if delta == 0 {
-                        continue;
-                    };
-                    tx.stop_codon = Some(cds_end + delta);
-                    let exon = gb
-                        .exons
-                        .iter_mut()
-                        .max_by_key(|g| g.alt_cds_end_i)
-                        .expect("No exons found during fix_cds");
-                    exon.alt_cds_end_i += delta;
-                    exon.cigar.push_str(&format!("{}I", delta));
-                    self.fixes
-                        .entry(Identifier::Transcript(
-                            TranscriptId::try_new(&tx.id).unwrap(),
-                        ))
-                        .or_default()
-                        .insert(Fix::Cds);
-                }
-            });
-    }
-
-    fn gene_name(&self, id: &Identifier) -> Option<String> {
-        match id {
-            Identifier::Gene(gene_id) => self
-                .gene_id_to_gene
-                .get(gene_id)
-                .and_then(|gene| gene.gene_symbol.clone()),
-            Identifier::Transcript(tx_id) => self
-                .transcript_id_to_transcript
-                .get(tx_id)
-                .and_then(|tx| tx.gene_name.clone()),
-        }
-    }
-
-    fn tags(&self, id: &Identifier) -> Option<Vec<Tag>> {
-        match id {
-            Identifier::Transcript(tx_id) => {
-                self.transcript_id_to_transcript.get(tx_id).map(|tx| {
-                    tx.genome_builds
-                        .values()
-                        .flat_map(|gb| gb.tag.clone())
-                        .flatten()
-                        .sorted_unstable()
-                        .dedup()
-                        .collect()
-                })
-            }
-            Identifier::Gene(gene_id) => {
-                self.gene_id_to_transcript_ids.get(gene_id).map(|tx_ids| {
-                    tx_ids
-                        .iter()
-                        .flat_map(|tx_id| {
-                            self.tags(&Identifier::Transcript(tx_id.clone()))
-                                .unwrap_or_default()
-                        })
-                        .sorted_unstable()
-                        .dedup()
-                        .collect()
-                })
-            }
-        }
-    }
-
-    fn discard(&mut self, remove: bool) -> Result<(), Error> {
-        let (n_transcripts_pre, n_gene_ids_pre) = (
-            self.transcript_id_to_transcript.len(),
-            self.gene_id_to_transcript_ids.len(),
-        );
-        type Counter = HashMap<Reason, usize>;
-        #[derive(Debug, Clone, Copy, Display, PartialEq, Eq, Hash)]
-        enum Kind {
-            Transcript,
-            GeneId,
-        }
-        let mut count_groups = HashMap::<Kind, Counter>::new();
-        let kinds = [Kind::Transcript, Kind::GeneId];
-        for kind in &kinds {
-            let mut counter = Counter::new();
-            for r in Reason::all().iter() {
-                counter.entry(r).or_default();
-            }
-            count_groups.insert(*kind, counter);
-        }
-
-        for (id, reason) in self.discards.clone() {
-            if reason.is_empty() {
-                tracing::warn!("Empty discard reason for {:#?}", id);
-                continue;
-            }
-            self._discard_id(&id, reason, remove)?;
-        }
-
-        for (id, reason) in &self.discards {
-            match id {
-                Identifier::Transcript(_) => {
-                    for r in reason.iter() {
-                        *count_groups
-                            .entry(Kind::Transcript)
-                            .or_default()
-                            .entry(r)
-                            .or_default() += 1;
-                    }
-                }
-                Identifier::Gene(_) => {
-                    for r in reason.iter() {
-                        *count_groups
-                            .entry(Kind::GeneId)
-                            .or_default()
-                            .entry(r)
-                            .or_default() += 1;
-                    }
-                }
-            }
-        }
-        let (n_transcripts_post, n_gene_ids_post) = (
-            self.transcript_id_to_transcript.len(),
-            self.gene_id_to_transcript_ids.len(),
-        );
-        tracing::info!(
-            "Discarded {} transcripts and {} Gene IDs",
-            n_transcripts_pre.abs_diff(n_transcripts_post),
-            n_gene_ids_pre.abs_diff(n_gene_ids_post)
-        );
-        for kind in kinds {
-            tracing::info!(
-                "{}: {:?}",
-                kind,
-                Reason::all()
-                    .iter()
-                    .filter(|r| count_groups[&kind][r] > 0)
-                    .map(|r| format!("{:#?}: {}", r, count_groups[&kind][&r]))
-                    .join(", ")
-            );
-        }
-        Ok(())
-    }
-
-    fn _discard_id(
-        &mut self,
-        id: &Identifier,
-        reason: BitFlags<Reason>,
-        remove: bool,
-    ) -> Result<(), Error> {
-        *self.discards.entry(id.clone()).or_default() |= reason;
-        match id {
-            Identifier::Gene(gene_id) => {
-                let txs = if remove {
-                    self.gene_id_to_transcript_ids.remove(gene_id)
-                } else {
-                    self.gene_id_to_transcript_ids.get(gene_id).cloned()
-                };
-                if remove {
-                    self.gene_id_to_gene.remove(gene_id);
-                }
-                for tx_id in txs.unwrap_or_default() {
-                    *self
-                        .discards
-                        .entry(Identifier::Transcript(tx_id.clone()))
-                        .or_default() |= reason;
-                    if remove {
-                        let _transcript = self.transcript_id_to_transcript.remove(&tx_id);
-                    }
-                }
-            }
-            Identifier::Transcript(tx_id) => {
-                let transcript = if remove {
-                    self.transcript_id_to_transcript.remove(tx_id)
-                } else {
-                    self.transcript_id_to_transcript.get(tx_id).cloned()
-                };
-
-                if let Some(hgnc) = transcript.and_then(|t| t.hgnc) {
-                    let gene_id: GeneId = hgnc.parse().expect("Invalid GeneId");
-
-                    if remove {
-                        let txs = self.gene_id_to_transcript_ids.get_mut(&gene_id);
-                        if let Some(txs) = txs {
-                            txs.retain(|tx| tx != tx_id);
-                            if txs.is_empty() {
-                                *self
-                                    .discards
-                                    .entry(Identifier::Gene(gene_id.clone()))
-                                    .or_default() |= Reason::NoTranscriptLeft;
-                                self.gene_id_to_transcript_ids.remove(&gene_id);
-                                self.gene_id_to_gene.remove(&gene_id);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Ok(())
-    }
-
-    fn mark_discarded(&mut self, id: &Identifier, reason: BitFlags<Reason>) -> Result<(), Error> {
-        if reason.is_empty() {
-            panic!("Empty discard reason for {:#?}", id);
-        }
-        *self.discards.entry(id.clone()).or_default() |= reason;
-        Ok(())
-    }
-
-    /// For each transcript that has been discarded for whatever reason, propagate the reason to its
-    /// parent gene id entry *if* the gene id entry already exists and has a non-empty reason
-    /// (to avoid erroneously discarding gene id entries for a non-important reason).
-    fn propagate_discard_reasons(&mut self, _raw: &Self) -> Result<(), Error> {
-        // First check whether all transcripts of a gene have been marked as discarded.
-        for (gene_id, _) in self.gene_id_to_gene.iter() {
-            let tx_ids = self
-                .gene_id_to_transcript_ids
-                .get(gene_id)
-                .map(|v| v.as_slice())
-                .unwrap_or_default();
-            if !tx_ids.is_empty()
-                && tx_ids.iter().all(|tx_id| {
-                    self.discards
-                        .get(&Identifier::Transcript(tx_id.clone()))
-                        .is_some_and(|d| d.intersects(Reason::hard()))
-                })
-            {
-                *self
-                    .discards
-                    .entry(Identifier::Gene(gene_id.clone()))
-                    .or_default() |= Reason::NoTranscriptLeft;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-/// Group transcripts by release and version (sorted descending).
-fn group_transcripts_by_release_and_version<'a>(
-    txs: &[&'a Transcript],
-) -> HashMap<(String, String), Vec<(u32, &'a Transcript)>> {
-    let mut versioned = txs
-        .iter()
-        .flat_map(|tx| {
-            let releases = tx.genome_builds.keys().cloned().collect_vec();
-            let tx_id = TranscriptId::try_new(&tx.id).unwrap();
-            let (ac, version) = tx_id.split_version();
-            let ac = ac.to_string();
-            releases
-                .into_iter()
-                .map(move |r| ((r, ac.clone()), (version, *tx)))
-        })
-        .into_group_map();
-    versioned.values_mut().for_each(|txs| {
-        txs.sort_unstable_by_key(|(version, _)| Reverse(*version));
-    });
-    versioned
-}
-
-fn append_poly_a(seq: String, length: usize) -> String {
-    // Append poly-A for chrMT transcripts (which are from ENSEMBL).
-    // This also potentially fixes the stop codon.
-    let mut seq = seq.into_bytes();
-    seq.extend_from_slice(b"A".repeat(length).as_slice());
-    String::from_utf8(seq).expect("must be valid UTF-8")
-}
 
 fn txid_to_label(
     label_tsv_path: impl AsRef<Path>,
@@ -619,17 +67,6 @@ fn txid_to_label(
                 })
         })
         .collect()
-}
-pub(crate) fn read_cdot_json(path: impl AsRef<Path>) -> Result<cdot_models::Container, Error> {
-    Ok(if path.as_ref().extension().unwrap_or_default() == "gz" {
-        tracing::info!("(from gzip compressed file)");
-        serde_json::from_reader(std::io::BufReader::new(flate2::read::GzDecoder::new(
-            File::open(path)?,
-        )))?
-    } else {
-        tracing::info!("(from uncompressed file)");
-        serde_json::from_reader(std::io::BufReader::new(File::open(path)?))?
-    })
 }
 
 /// Load the annotations (JSON or GFF3).
@@ -880,11 +317,11 @@ pub mod test {
     use temp_testdir::TempDir;
 
     use crate::common::Args as CommonArgs;
-    use crate::db::create::TranscriptLoader;
     use crate::db::create::cdot::load_cdot;
     use crate::db::create::cli::Args;
     use crate::db::create::filter::filter_transcripts;
     use crate::db::create::models::GeneId;
+    use crate::db::create::models::TranscriptLoader;
     use crate::db::dump;
 
     use super::run;
